requirements.txt
aiohttp==3.6.2
aiosqlite==0.12.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
oauthlib==3.1.0
PySocks==1.7.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
